requirements.txt
# Core ML and Computer Vision
numpy==1.19.5
opencv-python==4.5.3.56
tensorflow-aarch64==2.14.0
ml-dtypes==0.2.0
mediapipe==0.8.9.1

# Data Processing and Visualization
pandas==1.5.3
matplotlib==3.4.2
scikit-learn==1.1.3

# Image Processing
Pillow==8.3.1

# Web Interface
flask==2.0.1
werkzeug==2.0.1

# Utilities
protobuf==3.17.3
h5py==3.3.0
pytz>=2024.1 